pyproject.toml
[project]
name = "bitcrawler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "aiohttp-socks>=0.10.1",
    "aiosqlite>=0.21.0",
    "py7zr>=1.0.0",
    "pycryptodome>=3.23.0",
    "python-dotenv>=1.1.1",
    "rarfile>=4.2",
]
